+++ app/api/auth/[...nextauth]/route.ts
@@ -1,40 +1,6 @@
-// pages/api/auth/[...nextauth].ts (or app/api/auth/[...nextauth]/route.ts)
-import NextAuth, { NextAuthOptions } from "next-auth";
-import { PrismaClient } from "@prisma/client";
-import { PrismaAdapter } from "@next-auth/prisma-adapter"; // Note different package
-import DiscordProvider from "next-auth/providers/discord";
-
-const prisma = new PrismaClient();
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    DiscordProvider({
-      clientId: process.env.AUTH_DISCORD_ID!,
-      clientSecret: process.env.AUTH_DISCORD_SECRET!,
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  callbacks: {
-    jwt: ({ token, user }) => {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-    session: ({ session, token }) => {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.id as string,
-        },
-      };
-    },
-  },
-};
+// app/api/auth/[...nextauth]/route.ts
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth.config";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
